agent/windows/requirements.txt
grpcio>=1.63.0
grpcio-tools>=1.63.0
protobuf>=5.27.0
pywin32>=306
